app/requirements.txt
Flask~=2.3.2
SQLAlchemy~=2.0.15
WTForms~=3.0.1
requests~=2.25.1
psycopg2-binary~=2.9.6
Flask-WTF~=1.1.1
